refactor: socket 설정 파일 shared로 이동
