+++ client/src/shared/api/socket.ts
